Code cleanups surrounding effects limits Echo completed

+++ src/Effects/Reverb.cpp
@@ -563,7 +563,7 @@ unsigned char Reverb::getpar(int npar)
 
 float Revlimit::getlimits(CommandBlock *getData)
 {
-    float value = getData->data.value;
+    int value = getData->data.value;
     int control = getData->data.control;
     int request = getData->data.type & 3; // clear upper bits
     int npart = getData->data.part;
